refactor: members -> message_count にリネーム
