--- src/db/model/members.ts
@@ -1,8 +1,0 @@
-export class Members {
-    user_id: string;
-    message_count: number;
-    constructor(user_id: $TSFixMe, message_count: $TSFixMe) {
-        this.user_id = user_id;
-        this.message_count = message_count;
-    }
-}
