+++ src/db/model/message_count.ts
@@ -1,0 +1,8 @@
+export class MessageCount {
+    user_id: string;
+    count: number;
+    constructor(user_id: $TSFixMe, count: $TSFixMe) {
+        this.user_id = user_id;
+        this.count = count;
+    }
+}
